requirements_pi.txt
requests
gpsd-py3
rpi-lgpio